Chane tests names to have ITest postfix for integration tests
