+++ src/test/java/org/killbill/billing/plugin/vertex/health/VertexHealthcheckTest.java
@@ -41,7 +41,7 @@ import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
-public class VertexHealthcheckUnitTest {
+public class VertexHealthcheckTest {
 
     @Mock
     private Tenant tenant;
